supply historic data to strategies and use numpy internally

+++ src/shoe_backtest/typing.py
@@ -1,6 +1,7 @@
 from dataclasses import dataclass
 from enum import Enum
 
+import numpy as np
 import pandas as pd
 
 
@@ -13,7 +14,7 @@ class Side(Enum):
 class Transaction:
     date: pd.Timestamp
     symbol: str
-    qty: float
+    qty: np.float64
     side: Side
 
 
@@ -22,6 +23,6 @@ class PyfolioTransaction:
     date: pd.Timestamp
     symbol: str
     sid: int
-    amount: float
-    price: float
-    txn_dollars: float
+    amount: np.float64
+    price: np.float64
+    txn_dollars: np.float64
